@@ -9,12 +9,10 @@ class MessageReceivedEvent extends Event
 {
     public const EVENT_NAME = 'mqtt.message.received.action';
 
-    private $message;
+    public function __construct(
+        private Message $message
+    ) {}
 
-    public function __construct(Message $message)
-    {
-        $this->message = $message;
-    }
     public function getMessage(): Message
     {
         return $this->message;
